requirements.txt
bcrypt==4.1.3
Bcrypt-Flask==1.0.2
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
EbookLib==0.18
email_validator==2.1.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.2.2
MarkupSafe==2.1.5
python-docx==1.1.2
python-dotenv==1.0.1
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.12.2
Werkzeug==3.0.3
WTForms==3.1.2
